Update protobuf dep, adapt native source and sink (#209)
This PR updates the protobuf dependency and adapts the source and
sink to its new model. Of particular interest is that we no longer
have to copy as much when using protobuf, both by default in the
implementation and in our own manipulation. There's still some
clone inefficiency in sink/native.rs as a result of the implementation
of tagmap.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>