Bump to 0.5.2
This release incoporates repairs to the firehose sink which,
subsequent to 0.4.12, was accidentally nerfed. That is, the sink
was no longer able to store delivered lines into its queue. This
release also updates to serde 0.9, making rust 1.15 the minimum
rust version that can build cernan. We have also updated to
protobuf 1.2, reducing some of the cloning that goes on around
the native protocol source/sink.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>