Housekeeping commits (#211)
* Spruce up Cargo.toml for publishing to crates

We should have cernan up on crates.io. This commit is intended to
make cernan a respectable part of that ecosystem.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Remove explicit bincode dep, update to hopper 0.2.1

This commit removes the explicit bincode dependency that cernan
had. This was a hold-over from before hopper was split out into
its own library and has now been cleaned up. We also update to
hopper 0.2.1 to resolve a build issue against hopper 0.2.0 related
to hopper's bincode 1.0.0-alpha1 dependency.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>