This commit makes config parse errors more clear
David Kleiner had feedback for us in private correspondence that it
was difficult to tell why a config file would fail to parse. This
was partially due to all the 'unwrap' present in the code. We've
subbed this for 'expect' but should probably also build a validation
task because that config file has gotten pretty complicated.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>