Round out the flush-intervals work
This commit pushes the flush intervals work that Ivan's been doing
over the edge to completion. In particular, we now support:

  * per-sink configurable flush_intervals, plus a global default
  * guaranteed receive-once timer ticks

The later is done by including an every-increasing value into
TimerFlushes which allow duplicates -- because of paths in routing --
to be scrubbed out and ignored.

An example configuration is added at examples/configs/counting-example.toml
to let users fiddle with flush intervals. keep_count.lua has been
promoted to help in that effort.

This resolves #196.

This resolves #74.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>