Introduce payload clearing functions (#220)
* Introduce payload clearing functions

In conversation with @doubleyou I've learned that there's a use-case
for a filter that reads in a metric and then blanks the contents of
the payload. This is especially useful for filters that compute
analytics but want to dead-end streams of incoming metrics
excepting those that have been computed.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Correct misnaming

In my last commit I broke the naming scheme of lua internal functions,
working over pluralities. That is, we now have lua_clear_metrics | logs
as compared to lua_clear_metric | log.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>