Flush interval3 (#213)
* Have index in FlushTimer event

* Fix test to use TimerFlush with index

* ignore tags and vscode

* Turn off debugging prints

* will not introduce a new global constant here

* Remove trailing whitespaces

* Apply diff to config

* Apply other commits from flush-interval

* Last part

* Fix problems

* compiles

* flush interval in all the sinks

* removed flush_interval from null sink

* Round out the flush-intervals work

This commit pushes the flush intervals work that Ivan's been doing
over the edge to completion. In particular, we now support:

  * per-sink configurable flush_intervals, plus a global default
  * guaranteed receive-once timer ticks

The later is done by including an every-increasing value into
TimerFlushes which allow duplicates -- because of paths in routing --
to be scrubbed out and ignored.

An example configuration is added at examples/configs/counting-example.toml
to let users fiddle with flush intervals. keep_count.lua has been
promoted to help in that effort.

This resolves #196.

This resolves #74.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Remove the configurable interval from FlushTimer

It occurred to me that explaining the difference of the global
flush-timer as anything other than a default for sinks was
confusing and that I might as well make FlushTimer tick once
per second, hard-coded. Turns out that had been done by Ivan already
so I just kindly smoothed things out.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>