Allow lua filters to require other lua modules
This commit add explicit support for requiring modules in cernan
programmable filters, pegged by default to script-directory. We
make no attempts to disallow further modification of package.path
and the pegging to script-directory is meant to give the filter
programmer access to said directory without needing to know
exactly how cernan has been configured.

This is part of a larger work to support JSON as an input to filters.
See #229 for more details.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>