Introduce the 'internal' source (#235)
* Introduce the 'internal' source

This new source allows cernan to self-telemeter. It is not possible
for modules to report telemetry 'up' to be distributed through
the filter topology as operators see fit. This interface is
intentionally minimal and can be extended as need arises.

Closes #212

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Address formatting feedback

This commit is a combination of addressing outstanding cargo fmt
flubs and whitespace policing by John.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>