Make sampling inclusive of all statsd types (#236)
* Make sampling inclusive of all statsd types

@pulltab reported a metric being flagged as BAD METRIC in his systems
logs which looked roughly like so:

    important-db-latency:1.001|ms|@0.25

The trick was that the statsd protocol implementation understood samples
for gauges and counters but not histograms or timers. This commit
corrects that problem and applies sampling equally to all aggregation
types.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Make statsd quickcheck protocol test more comprehensive

The previous verison of this test merely asserted that the generated
lines would parse. This commit includes checks to confirm that, in fact,
the parsed Telemetry objects make good sense.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>