Introduce better image building
This commit introduces a work-in-progress update to the image
building process for cernan. The ambition is to build an image
for each new release and in-progress cuts of master as well. The
busybox:glibc image is used as a base for production -- see
Dockerfile.prod -- and postmates/postmates-rust.

The image built in this commit is not quite correct but it's
a start. It also depends on closed-source postmates-rust, a thing
which we will need to open-source.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>