Protect against duplicate timestamps in prometheus sink
Prometheus has a fiddly protocol. As a reporter you have to be
careful not to timestamp the same which is challenging when you
are pulled from, not pushing. What we do -- as explained in
comment -- is partition all our aggregations and only report
when they're in the past. This is not perfect -- you know, because
clocks -- but it's better than naught.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>