Use raw headers, correct partition inequality
It turns out we were missing metrics for three reasons:

  * MetricFamily in protobuf did not have its name set,
  * the partition for 'not ripe' metrics was incorrect and
  * the content-type header was not specific enough.

To the last point we now just plonk strings down into the header.
Fancy typed enum headers are useful but, uh, this is just a static
string. I now confirm that in a local setup I do, in fact, see
the metrics I expect in prometheus.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>