Add 'fields' to LogLine, filter manipulation of such (#241)
* Add 'fields' to LogLine, filter manipulation of such

We need to start parsing structured information out of LogLines and this
work enables it. What we've done is plonk an additional k/v structure
inside the LogLine that filters have access to, are able to manipulate.
We demonstrate this under test by pulling information from a json payload.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Teach firehose how to understand LogLine fields

This commit extends the firehose sink so that it is able to place
the fields k/v into the payload proper. Firehose has no metadata
concept so the same method is used for fields as for tags.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>