Cope with prometheus header behaviour (#239)
* Cope with prometheus header behaviour

It turns out that prometheus is a little more lax in practice with
regard to content-type headers than its documentation promises. This
commit meets prometheus where it is, emitting text/plain if there's
no header and not emitting protobuf by default.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Protect against duplicate timestamps in prometheus sink

Prometheus has a fiddly protocol. As a reporter you have to be
careful not to timestamp the same which is challenging when you
are pulled from, not pushing. What we do -- as explained in
comment -- is partition all our aggregations and only report
when they're in the past. This is not perfect -- you know, because
clocks -- but it's better than naught.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Ensure prometheus text response has correct tags, handle accept header

We previously did not understand that prometheus would send an accept
header and we ought to reply with content-type. As you can guess,
I am not up on my HTTP. We now properly negotiate types with prometheus.

As well, prometheus text response now correctly encodes outgoing tag
information.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Use raw headers, correct partition inequality

It turns out we were missing metrics for three reasons:

  * MetricFamily in protobuf did not have its name set,
  * the partition for 'not ripe' metrics was incorrect and
  * the content-type header was not specific enough.

To the last point we now just plonk strings down into the header.
Fancy typed enum headers are useful but, uh, this is just a static
string. I now confirm that in a local setup I do, in fact, see
the metrics I expect in prometheus.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Extend sanitization effort in prometheus sink

This commit improves the random nonsense that will populate an
Arbitrary Telemetry by prefering to use gen_iter::<char> rather
than gen_ascii_bytes. This allows us to QC the sanitization function
of the prometheus sink which is, uh, unfortunately a little more
copy'y than I'd like. I don't know String well enough to know if
there's a better way to do the manipulation in-place while looping
around.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>