Build a special-purpose aggregator for Prometheus
The constraints placed by Prometheus on a reporter are non-trivial,
especially with regard to reports through time. This commit introduces
a new aggregator for Prometheus and a battery of quickcheck tests to
demonstrate singleton reports of time-series / timestamp, re-introduction
of telemetry after report failure and documentation of such.

We have to discover Prometheus' constraints, in part, and the model
presented in this commit may diverge from Prometheus' actual demands.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>