Eject the timestamp concept from Prometheus sink
Inspired by a comment in the documentation that says that prometheus
reporters should not issue their own timestamps -- compounded by odd
behaviour we found when doing so -- we've elected to stop sending
timestamps. Now, in so doing we've had to drop the time notion from
the aggregation, which has simplified the aggregation. We now only
care for the name and tagmaps of a time series.

This does mean that 'binning' as such is not obeyable in the
Prometheus sink. Fortunately the CKMS structure means we will not
be overloaded with points should prometheus hang up for a while.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>