Build a special-purpose aggregator for Prometheus (#242)
* Build a special-purpose aggregator for Prometheus

The constraints placed by Prometheus on a reporter are non-trivial,
especially with regard to reports through time. This commit introduces
a new aggregator for Prometheus and a battery of quickcheck tests to
demonstrate singleton reports of time-series / timestamp, re-introduction
of telemetry after report failure and documentation of such.

We have to discover Prometheus' constraints, in part, and the model
presented in this commit may diverge from Prometheus' actual demands.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Eject the timestamp concept from Prometheus sink

Inspired by a comment in the documentation that says that prometheus
reporters should not issue their own timestamps -- compounded by odd
behaviour we found when doing so -- we've elected to stop sending
timestamps. Now, in so doing we've had to drop the time notion from
the aggregation, which has simplified the aggregation. We now only
care for the name and tagmaps of a time series.

This does mean that 'binning' as such is not obeyable in the
Prometheus sink. Fortunately the CKMS structure means we will not
be overloaded with points should prometheus hang up for a while.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Clear up formatting issues

The previous commit was bounced by CI on account of its formatting
flubs. This commit fixes said flubs. Someday there will be a robot
to do this and it will be glorious.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>