Add telemetry to InfluxDB sink, chunk
This commit does not attempt to send all the points stored in
the InfluxDB buffer at once. It opens a socket and then sends
multiple payloads. That socket is re-used on error, though it's
not clear if this is right to do or what errors might be bubbled
up. Further, we add telemetry around reporting to InfluxDB.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>