Travis test fixes (#250)
* Remove format check from travis

Rustfmt is changing too quickly for this to be a valuable check.
It was worth the experiment.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Reduce QC test runtime

The travis tests are taking up to twenty minutes to run. This is
driven by QC tests searching for just the right setup in the
prometheus sink. We improve the probability of this search by
dropping the max length of a Telemetry string. We reduce the
runtime cost by reducing the number of Telemetry items we must
make.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>