Correct read / swap issue in file source
This commit corrects an issue where a file source would only function
in older rust releases. The culprit, as I understand it, is that
BufReader will return reads of size zero, which the glob loops was
not able to cope with. This is now treated as an error condition.
Further, we keep track of our last known-good offset inside of the
FileWatch, allowing us to back up as needed, rather than relying
on the unspecified behaviour of BufReader.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>