Steps toward InfluxDB in prod (#248)
* Update deps for 'cargo publish'

For reasons that, uh, aren't entirely clear to me the whole build
process went well but 'cargo publish' crapped out. This commit
updates the dependencies of cernan so that 0.5.9 can be published.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Move InfluxDB sink toward production status

With the InfluxDB sink we're doing something different from the
other sinks: no aggregation. InfluxDB is a proper timeseries database
and has a rich query language. There's no need for cernan to do
pre-aggregation. InfluxDB further has nanosecond timestamps so
timestamp_ns has been added to the Telemetry struct to accomidate
this.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Add telemetry to InfluxDB sink, chunk

This commit does not attempt to send all the points stored in
the InfluxDB buffer at once. It opens a socket and then sends
multiple payloads. That socket is re-used on error, though it's
not clear if this is right to do or what errors might be bubbled
up. Further, we add telemetry around reporting to InfluxDB.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Correct CKMS comparison in test

Because of floating point comparison sometimes our friend
buckets::test_all_insertions would fail. This was detected only
in the beta compiler channel, so I'm real glad we test across
all channels.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Relax patch target from 100% to 90%

For reasons that fully escape me some runs kcov will sigsev on
the passed cernan test binary. This is... not great. It's also
not something that I can reproduce.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>