Shift InfluxDB to use the http/s interface, not UDP
This commit modifies the InfluxDB sink so that it uses the HTTP/S
interface of InfluxDB. Previously we were using UDP which is nice
and all but requires special support from InfluxDB operators and
is lossy, with the technique that we were using.

Best to default to reliable and slow than unreliable and fast, I
guess.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>