Shift InfluxDB to use the http/s interface, not UDP (#253)
* Shift InfluxDB to use the http/s interface, not UDP

This commit modifies the InfluxDB sink so that it uses the HTTP/S
interface of InfluxDB. Previously we were using UDP which is nice
and all but requires special support from InfluxDB operators and
is lossy, with the technique that we were using.

Best to default to reliable and slow than unreliable and fast, I
guess.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Update documentation, change how delivery backoffs work

This commit changes the mechanism by which delivery backoffs are
computed by removing the cap on delivery_attempts. The justification
for this is given in comments: namely, we wish to avoid swamping a
failing DB.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Ensure that InfluxDB connections are kept open as long as possible

This commit makes better use of Hyper by understanding that 1. the
client is automatically pooled with five connections and that 2. the
only thing we need to do to make the pool more efficient is add a
keep-alive header. This is now done. Look at us, so fancy.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>