Update docker build (#238)
* Introduce better image building

This commit introduces a work-in-progress update to the image
building process for cernan. The ambition is to build an image
for each new release and in-progress cuts of master as well. The
busybox:glibc image is used as a base for production -- see
Dockerfile.prod -- and postmates/postmates-rust.

The image built in this commit is not quite correct but it's
a start. It also depends on closed-source postmates-rust, a thing
which we will need to open-source.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Shuffle dockerfile images around

We shouldn't build a docker image in a docker image. What we're
going to try and do is build an image and then 'docker cp' contents
out.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Achieve fully static build of cernan

This commit makes it possible to build a fully static binary of
cernan -- linked with musl -- and placed inside of a busybox
image.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Basic confd default cfg example in the container

* Route the internal source in basic image toml

The internal source is often very useful to have a debugging
environment. This commit enables it and routes it properly to
sinks.console.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Re-up Cargo.lock / toml

We had a build failure around my fork of rust-lua53 on account of,
well, I'm not sure. A 'cargo update' corrected the problem.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* No need for AWS_REGION, other cleanup bits

* Attempt to use libressl instead of openssl

 - requires https://github.com/postmates/postmates-rust/pull/2

* Build a dynamic cernan on alpine:edge + testing

This commit uses the stable rustc compiler in alpine testing and
targets the same for production. The resulting setup _should_
work where openssl is needed on account of us not linking it
statically.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Remove accidental comments

The commented out code here really makes it hard to use build-container,
on account of how it disables the important bits. Oops

Signed-off-by: Brian L. Troutwine <blt@postmates.com>