Ensure that 'git' is in the testing docker (#256)
* Ensure that 'git' is in the testing docker

It turns out it'd be very useful to have git available for some
docker work. So! We're going to make it available in the test
environment.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Introduce kcov for coverage runs

We want to have coverage checking via codecov. This commit includes
kcov and friends into the test image.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Reduce build-times in test main path

This commit removes the installation of rustfmt, clippy, fuzz. These
are _not_ needed to run normal tests and only slow down the build.
CI can add them as required.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Trim kcov from the test image

The coverage setup for cernan has never been super solid. Apparently
we were ignoring our integration tests and fuzzing never would have
fit. As such we drop it from the test image in the interests of making
a comprehensive solution later.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>