Tests now pass
With this commit tests are now passing again. It's nice! There's
been some fiddling in bin/cernan to reduce the copying we once did.
I guess when I wrote that I didn't rightly know about mem::replace
but the pattern was obvious on read-through.

There's been fiddling all over the system because of rustfmt. It's
good to have rustfmt turned on by default in _some_ sense but it
does make the commits blow out into nonsense.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>