Require --config to be passed, resolve wrap sub crash
The configuration situation for cernan is... odd. Used to be there
were a whole host of flags you could pass to get behaviour but we
stopped using those and no one wanted any others. Now that's
more strictly enforced: you have to pass a configuration.

I also corrected a crashing bug that can happen if you pass a
blank config.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>