Update dependencies, configuration  (#259)
* WIP update to config.rs

This is a big change. The ambition is to enable feature flags and
update our serde using libraries to all 1.0, allowing us to build
without having to do wacky pinning.

Right now everything is in pieces. That'll get fixed.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Now passing configuration tests

With this commit we're now passing our configuration tests. There's
been nothing done to improve the rusoto situation and our sources
are still split out into individual fields. I _think_ if we combine
them then we'll have and easier time enabling feature flags.

I dunno.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Tests now pass

With this commit tests are now passing again. It's nice! There's
been some fiddling in bin/cernan to reduce the copying we once did.
I guess when I wrote that I didn't rightly know about mem::replace
but the pattern was obvious on read-through.

There's been fiddling all over the system because of rustfmt. It's
good to have rustfmt turned on by default in _some_ sense but it
does make the commits blow out into nonsense.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Take clippy's suggestions

With this commit clippy is now silent on cernan. Wee! This was a
nice surprise. This resolves #178.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Remove print statements, bump to 0.6.0-pre

This commit removes some stray print statements left in the codebase.
Additionally, we are now targeting a release of 0.6.0.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Fiddle with the example configuration files

This commit fiddles with the example configuration files on account
of how we've changed things around the native sink. I can't seem
to cause a failure when integration testing so, uh, I think we're
good here.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Further clippy suggestions

This commit includes further clippy suggestions. I was wrong. We are
not clippy-quiet on account of the explict loops. I _believe_ we can
fix these because my iterator is, uh, goofy.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Require --config to be passed, resolve wrap sub crash

The configuration situation for cernan is... odd. Used to be there
were a whole host of flags you could pass to get behaviour but we
stopped using those and no one wanted any others. Now that's
more strictly enforced: you have to pass a configuration.

I also corrected a crashing bug that can happen if you pass a
blank config.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Update README to reflect --config changes

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Update dependencies

For some reason the lua dependency has started to fail on account
of readline. This is... weird. It's weird since we're stuck using
my own fork of the lua dep that specifically removes the use of
readline from lua.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Update test docker environment to run clippy

This was something I had intended but never actually did. Oops!
The output for clippy is not easy for a machine to parse but that's
okay. We've got something going for us, anyway.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>