Fix the name mangling issue with self-telemetry
When the internal metric source was created we implicitly encoded
a graphite-like scheme for metric naming, which, since cernan intends
to be agnostic to sink, was not ideal. This commit corrects that
mistake, allowing callers to report_telemetry to expand their
use to report_full_telemetry, a function which admits metadata and
changing the aggregation type.

This resolves #262

Signed-off-by: Brian L. Troutwine <blt@postmates.com>