Correct our use of Arbitrary
Because I didn't rightly misunderstand how to write these out
when cernan started there was a pattern of Rand + Arbitrary through
the codebase. We only needed Arbitrary. This is now corrected. Little
has changed but the code is somewhat cleaner as a result.

We've dropped the explicit dependency on rand.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>