Last over-the-line work for direct elasticsearch
As of this commit I am able to push into a local docker ES.

    > docker pull docker.elastic.co/elasticsearch/elasticsearch:5.4.1
    > docker run -p 9200:9200 -e "xpack.security.enabled=false" -e "http.host=0.0.0.0" -e "transport.host=127.0.0.1" docker.elastic.co/elasticsearch/elasticsearch:5.4.1

There's no authentication cernan-side because the initial consumer
is not going to make use of auth. We'll have to figure something out
here on account of putting passwords into a config file on disk is
probably not the best.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>