Update the way error handling works some
This commit reduces some of the logging we have internal to cernan
as well as the duplication around errors. It's still a mystery to
me which errors in ES are recoverable and which are not. Or, rather,
it's not a myster but so far as I can tell it seems that every error
_should_ eventually be recoverable in the way we have this coded.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>