Direct elasticsearch integration (#268)
* Initial introduction of direct Elasticsearch

This commit introduces direct Elasticsearch integration. I make
no claim that it works, only that it compiles.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Last over-the-line work for direct elasticsearch

As of this commit I am able to push into a local docker ES.

    > docker pull docker.elastic.co/elasticsearch/elasticsearch:5.4.1
    > docker run -p 9200:9200 -e "xpack.security.enabled=false" -e "http.host=0.0.0.0" -e "transport.host=127.0.0.1" docker.elastic.co/elasticsearch/elasticsearch:5.4.1

There's no authentication cernan-side because the initial consumer
is not going to make use of auth. We'll have to figure something out
here on account of putting passwords into a config file on disk is
probably not the best.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Formatting updates

This commit does not change the workings of the elasticsearch
module, only improves its formatting.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Don't abort on panic

This setting can't be combined with some of our deps when we compile
in release mode. Oops

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Configurable prefix for the ES index

This commit makes it possible to add a prefix to the ES index and
do so as a matter of configuration.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Don't clear on every loop

Because we're popping from a buffer there's no need for use to
go nuts and clear the buffer after we submit one point.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Add trace messages into ES sink

It's not clear to me what's up here. I need more info.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Add more trace messaging into the sink

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Make types, mappings etc

I didn't realize at the start that ES had to have this little dance
done. Now I know.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Change how we index docs

This commit updates how we go about indexing to be inline with the
ES documentation. It should be accurate.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Add debugging / loggin info

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Update the way error handling works some

This commit reduces some of the logging we have internal to cernan
as well as the duplication around errors. It's still a mystery to
me which errors in ES are recoverable and which are not. Or, rather,
it's not a myster but so far as I can tell it seems that every error
_should_ eventually be recoverable in the way we have this coded.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Increase testing for elasticsearch configuration

This commit adds a test so that we have minimal confidence in the
configuration of the elasticsearch sink.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>