Address Elasticsearch 0.6.1 issues (#277)
* Add tags as nested objects to ES payload

This commit adds a Metadata object as a nested object to the main
Payload. It's hoped that we'll be able to index over tags and
associate them with payloads. Something of an experimental commit.

Addresses #276

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Build explicit JSON documents

The use-case for this sink is to drop into Elasticsearch and then
read them through kibana. Kibana doesn't have the ability to cope
with nested objects so we need to cram the metadata into the body
of each document.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Correctly format timestamps as ISO 8601 / RFC 3339

The auto-conversion left out all the nice colons and dashes that
the above format requires. Oops

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Add additional logging

We're still not getting tags through. Unsure what's up there but
we need logging.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Assert that the inner buffer is not empty

If the inner ES buffer is empty we really shouldn't be building
a body. Something bad has happened.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Log the ES body in debug mode

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Be sure to populate tags into File source

Signed-off-by: Brian L. Troutwine <blt@postmates.com>