Make buckets a per-telemetry iterator
The bucket abstraction has previously done its iteration per-value
array, leaking information about how buckets are implemented. This
was done because I didn't really understand how to make an effective
iterator.

The goal of this commit is to set us better for implementing zero-padding
in the wavefront sink. You can see some of that work is already taking
place, notably with the set of seen telemetry, allowing us to fill
in patches.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>