Further work on the padding iterator
Here's the deal: we have to take ownership of the stream of Telemetry
coming out of the buckets. This is counter to what we've normally done
because we've just worked by reference and then later cleared. Problem
is we gotta add new stuff into the stream and that has to be allocated
_somewhere_.

I dunno. Maybe if I wrap everything up in RCs

Signed-off-by: Brian L. Troutwine <blt@postmates.com>