Add commenting into the padding
This commit adds commenting so that the padding routine can be
understood. It's a non-trivial bit of code and I know that I won't
recall quite how it works in the Glorious Future without a little
help.

What remains now is to emit a point across flushes if a point has
not come in recently. We have the seen-set available for that,
kind of. The problem is we don't store enough information to
reconstruct the telemetry in the even of gap.

I'll drop that stuff in the next commit and make the pull request
clean.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>