Keep sockets in native / wavefront
The previous implementation of these sinks created and destroyed sockets
on each flush. There is no good reason for this any longer. We now take
the approach of preserving the socket only so long as we're able to send
information through it, otherwise we destroy the socket and get a new one.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>