Wavefront zero padding (#283)
* Make buckets a per-telemetry iterator

The bucket abstraction has previously done its iteration per-value
array, leaking information about how buckets are implemented. This
was done because I didn't really understand how to make an effective
iterator.

The goal of this commit is to set us better for implementing zero-padding
in the wavefront sink. You can see some of that work is already taking
place, notably with the set of seen telemetry, allowing us to fill
in patches.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Further work on the padding iterator

Here's the deal: we have to take ownership of the stream of Telemetry
coming out of the buckets. This is counter to what we've normally done
because we've just worked by reference and then later cleared. Problem
is we gotta add new stuff into the stream and that has to be allocated
_somewhere_.

I dunno. Maybe if I wrap everything up in RCs

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Revamp Buckets iterator, passing tests

This is the first commit in the series with passing tests. We aren't
checking to see if the padding works yet but, based on how Wavefront
is passing it appears that we're doing fine when there's no padding
to be done.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Intermediate step

This commit is an intermediate step. The iterator has gotten much
more complicated and this is due to gradually coping for edge cases
in the padding iterator. There's a simpler approach here, I'm sure,
but it's not appearing to me right now.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* wavefront::test::test_no_unpadded_gaps now passing

It's a nice feeling when a difficult quickcheck test finally passes.
There is one more property to verify before wrapping this chain
up into a pull-request but we're closer. Those properties done and
to do are:

  [x] never a gap without two zeros on either side
  [ ] never fewer non-zero points

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Clean up some printlns, introduce last property test

This commit adds the property test to check that we don't drop
any non-zero values when we do padding. I've also cleaned up a
load of printlns that were useful during debugging. As well, the
manual test is removed as this is not needed for debugging failing
QC tests.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Add commenting into the padding

This commit adds commenting so that the padding routine can be
understood. It's a non-trivial bit of code and I know that I won't
recall quite how it works in the Glorious Future without a little
help.

What remains now is to emit a point across flushes if a point has
not come in recently. We have the seen-set available for that,
kind of. The problem is we don't store enough information to
reconstruct the telemetry in the even of gap.

I'll drop that stuff in the next commit and make the pull request
clean.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Remove 'seen_telemetry' concept from Wavefront sink

This commit follows up on my promise to remove the seen_telemetry
concept from the wavefront sink. It, uh, does exactly that.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Remove time argument for wavefront::format_stats

The time value that was being passed here was not used. This was
flagged by Tom Santero during review and the argument is now removed.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Uncomment println in wavefront test function

The commenting of this println was accidental and was flagged by
Tom Santero during review. This is corrected.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>