Elide excess zeros in wavefront flushes
This commit makes it so we don't ship overly many zeros to Wavefront.
We do this by stripping _all_ zero-valued Telemetry out of the stream
and then padding that with zeros as needed. This is 2*n loops through
the stream but we can fiddle with the implementation to cause fusion
_or_ just wait for the compiler to get sufficiently smart.

I vote wait.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>