Implement 'sustain' for wavefront sink
This commit implements a 'sustain' notion for wavefront. That is,
time series will continue on after the last point for 'sustain' flush
intervals. The implementation is, uh, a little pokey because of the
clones but it can't be helped without a fundamental rethink of how
we maintain Telemetry in-memory in the Wavefront sink. Nothing came
to mind though.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>